requirements.txt
aiohttp==3.8.1
aiosignal==1.3.1
appdirs==1.4.4
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
braceexpand==0.1.7
bs4==0.0.1
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
cloudscraper==1.2.68
colorama==0.4.6
cssselect==1.2.0
dnspython==2.2.1
easy-pil==0.1.9
emoji==2.2.0
fake-useragent==1.1.1
filelock==3.8.0
frozenlist==1.3.3
google-api-core==2.10.2
google-api-python-client==2.66.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.57.0
grapheme==0.6.0
html5lib==1.1
httplib2==0.21.0
idna==3.4
import-expression==1.1.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
lxml==4.9.2
markdown-it-py==2.1.0
MarkupSafe==2.1.2
mdurl==0.1.2
motor==3.1.1
multidict==6.0.2
parse==1.19.0
Pillow==8.4.0
platformdirs==2.5.4
prettytable==3.5.0
protobuf==4.21.9
py-cord
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyee==8.2.2
pymongo==4.3.2
pyparsing==3.0.9
pyppeteer==1.0.2
pyppeteer-stealth==2.7.4
pyquery==2.0.0
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
requests-html==0.10.0
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
tqdm==4.64.1
typing_extensions==4.4.0
uritemplate==4.1.1
uritools==4.0.0
urlextract==1.7.1
urllib3==1.26.12
w3lib==2.1.1
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.4
yarl==1.8.1
zipp==3.12.0
